♻️ 83KGYN hooks: decompose pre-push helpers
