+++ .agentplane/tasks/202605290532-83KGYN/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 83KGYN task: Pre-push hook decomposition [202605290532-83KGYN]
