+++ .agentplane/tasks/202605290532-83KGYN/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290532-83KGYN/pre-push-hook-decomposition",
+  "created_at": "2026-05-29T05:33:16.171Z",
+  "diffstat_sha256": "sha256:52af8be2e30417854eca87d136d0b8bab70c5c71a82312d64ec1ee876898b0d9",
+  "last_verified_at": "2026-05-29T05:38:25.859Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290532-83KGYN",
+  "updated_at": "2026-05-29T05:33:16.171Z",
+  "verify": {
+    "status": "pass"
+  }
+}
